+++ backend/src/main/java/com/debede/backend/repository/TiendaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TiendaRepository extends JpaRepository<Tienda, Long> {
+public interface TiendaRepository extends JpaRepository<Tienda, Integer> {
 
 }
